requirements.txt
# Core Data & Analysis
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28

# Technical Analysis
pandas-ta>=0.3.14b0
scipy>=1.10.0

# API & Data
requests>=2.31.0
aiohttp>=3.9.0

# Caching
diskcache>=5.6.0

# UI
streamlit>=1.28.0
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
